pyproject.toml
[tool.poetry]
name = "perfi"
version = "0.1.0"
description = "defi portfolio analytics"
authors = ["Your Name <you@example.com>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
psutil = "^5.9.0"
httpx = "^0.22.0"
rootpath = "^0.1.1"
arrow = "^1.2.2"
jsonpickle = "^2.1.0"
XlsxWriter = "^3.0.3"
pydantic = "^1.9.0"
fastapi = "^0.75.1"
typer = "^0.4.1"
tqdm = "^4.64.0"
openpyxl = "^3.0.9"
beautifulsoup4 = "^4.10.0"
lxml = "^4.8.0"
prettytable = "^3.2.0"
rich = "^12.1.0"
click = "^8.1.2"
sttable = "^0.0.1"
pytz = "^2022.1"
uvicorn = "^0.17.6"
protobuf = "3.20.0"
siwe = "^1.0.1"
itsdangerous = "^2.1.2"
coinaddrvalidator = "^1.1.3"
python-multipart = "^0.0.5"
matplotlib = "^3.5.2"
pandas = "^1.4.2"
squarify = "^0.4.3"
plotly = "^5.8.0"
tabulate = "^0.8.9"
"delegator.py" = "^0.1.1"
kaleido = "0.2.1"
colorama = "^0.4.4"
numpy = "^1.22.4"
scipy = "^1.8.1"
CurrencyConverter = "^0.16.12"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pre-commit = "^2.18.1"
devtools = "^0.8.0"
sttable = "^0.0.1"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
